refactor(insights): extract types, utils, and MathSelector from ActionFilterRow
ActionFilterRow.tsx was 1,446 lines with 6+ concerns in one file. This
extracts types, utility functions, and the MathSelector component into
focused modules while preserving backward-compatible re-exports.

- types.ts: MathAvailability, ActionFilterRowProps, MathSelectorProps
- actionFilterRowUtils.ts: getValue, taxonomicFilterGroupTypeToEntityType
- mathUtils.ts: type guards, getDefaultPropertyMathType, histogram set
- MathSelector.tsx: thin wrapper component
- useMathSelectorOptions.tsx: 390-line hook for math dropdown options

ActionFilterRow.tsx goes from 1,446 to 874 lines. All 28+ existing
import sites continue to work via re-exports.
